Summarization-Backend/SummarizationBackendApi/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
requests==2.31.0
OpenAI==0.27.8
langchain==0.0.247
pypdf==3.14.0
tiktoken== 0.4.0
docx2txt==0.8
python-decouple==3.8
